+++ scripts/BuildAll.ps1
@@ -1,7 +1,8 @@
 param()
 
 & "$PSScriptRoot\BuildUnpacked.ps1"
-& "$PSScriptRoot\BuildWin32.ps1"
-& "$PSScriptRoot\BuildWin64.ps1"
+& "$PSScriptRoot\BuildWinX86.ps1"
+& "$PSScriptRoot\BuildWinX64.ps1"
+& "$PSScriptRoot\BuildWinArm64.ps1"
 & "$PSScriptRoot\BuildMacOS.ps1"
 & "$PSScriptRoot\BuildLinux.ps1"